\name{Ringo-internal}
\alias{clusters}
\alias{panel.scatter}
\alias{panel.cor}
\alias{merge.RGList}
\alias{takeMeanOverGroups}
\alias{as.data.frame.cherList}
\alias{splitAndSimplify}
\alias{readNimblegenHeader}
\alias{readNgIntensitiesTxt}
\alias{normalizeBetweenArraysVSN}
\alias{relateChers2}
\alias{whichCsr}
\alias{oneChannelVSN}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Internal Ringo functions}
\description{
  Called internally by other \code{Ringo} functions.
  Normally need not be called by the user.
}
\author{Matt Ritchie, Tammo Krueger, Joern Toedling}
\keyword{internal}% at least one, from doc/KEYWORDS
