% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ringo.R
\docType{package}
\name{ringo}
\alias{ringo}
\title{ringo: Read STAR files into R}
\description{
The ringo package is for reading and manipulating STAR files like those generated by
Relion \url{https://www2.mrc-lmb.cam.ac.uk/relion/index.php?title=Main_Page}
into R.
}
