% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starr.R
\name{single_table_to_df}
\alias{single_table_to_df}
\title{Single Table to Tibble}
\usage{
single_table_to_df(z, clean_relion_names = TRUE)
}
\arguments{
\item{z}{A Vector of all the lines that make one STAR table}
}
\value{
A tibble
}
\description{
Turns a single table (as a vector of lines) from the STAR file into a tibble.
}
