\name{Ringo-internal}
\alias{clusters}
\alias{panel.scatter}
\alias{panel.cor}
\alias{merge.RGList}
\alias{takeMeanOverGroups}
\alias{as.data.frame.cherList}
\alias{splitAndSimplify}
\alias{readNimblegenHeader}
\alias{readNgIntensitiesTxt}
\alias{normalizeBetweenArraysVSN}
\alias{relateChers2}
\alias{oneChannelVSN}
\alias{exportCCData}
\alias{pair2xys}
\alias{alongChromTicks}
\alias{featureColors}
\alias{featureColors}
\alias{plotAlongChromLegend}
\alias{newVP}
\alias{plotOneChIPSample}
\alias{cbind2,ExpressionSet,ExpressionSet-method}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Internal Ringo functions}
\description{
  Called internally by other \code{Ringo} functions.
  Normally need not be called by the user.
}
\author{Matt Ritchie, Tammo Krueger, Wolfgang Huber, Joern Toedling}
\keyword{internal}% at least one, from doc/KEYWORDS
