\name{Ringo-internal}
\alias{clusters}
\alias{panel.scatter}
\alias{panel.cor}
\alias{merge.RGList}
\alias{takeMeanOverGroups}
\alias{addTypes}
\alias{typifyList}
\alias{getGeneToTranscriptMap}
\alias{getTranscriptToGeneMap}
\alias{combine.tables}
\alias{calcOverlapDfVsDf}
\alias{as.data.frame.peakList}
\alias{generatePeakList}
\alias{splitAndSimplify}
\alias{readNimblegenHeader}
\alias{readNgIntensitiesTxt}
\alias{normalizeBetweenArraysVSN}
\alias{relatePeaks2}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Internal Ringo functions}
\description{
  Called internally by other \code{Ringo} functions.
  Normally need not be called by the user.
}
\author{Matt Ritchie, Tammo Krueger, Joern Toedling}
\keyword{internal}% at least one, from doc/KEYWORDS
